OH MY GOD okay so the issue was it was running a ticker behaviour that was repeating the send, so changed that to wakerbehaviour, then also changed the reply on the student agent to a blockingMessage, then changed ALL the integers to longs because it wasn't liking my integers so now messages are actually being passed just for some reason the timetabler is sending multiple copies of the tutorial offers to each student? so i suppose that's next lol

+++ src/ModuleGeneration.java
@@ -50,13 +50,13 @@ public class ModuleGeneration
         for (int i = 0; i < tutorialGroups; i++) {
             var tut = new Tutorial();
             tut.setModuleId(moduleId);
-            tut.setCapacity(evenTutorialSize);
+            tut.setCapacity((long) evenTutorialSize);
             tutorials.add(tut);
             
         }
         
         if (numberOfEnrolledStudents % 2 != 0) {
-            tutorials.get(tutorialGroups - 1).setCapacity(evenTutorialSize + 1);
+            tutorials.get(tutorialGroups - 1).setCapacity((long) (evenTutorialSize + 1));
         }
         
         var module = new Module();
